requirements.txt
boto3==1.14.18
Flask==1.1.2
google-cloud-storage==1.29.0
jsonschema==3.2.0
